refactor: split modules and add robust shutdown
