+++ CHANGELOG.md
@@ -1,0 +1,8 @@
+# Changelog
+
+## [Unreleased]
+
+### Added
+- Initial module split for `ScreenTracker v2`.
+- Basic shutdown robustness with signal handling, window-close detection and UI watchdog.
+
