+++ rules/__init__.py
@@ -1,0 +1,2 @@
+"""Rule plug‑ins package."""
+
